Renamed the SubscriptionPlan folder in the modules directory and updated all import paths where the folder is used.
Renamed the SubscriptionPlan folder in the modules directory and updated all import paths where the folder is used.
